Updates from cldev-teameval
Yes, this is a very big commit. It integrates changes from multiple projects, including
contributions from the University of New South Wales and Blackboard Australia. It includes
the following improvements:

* Support for Team Evaluation plugin (UNSW)
* Automatic recalculation of calibrations scores (BB)
* Submit on behalf of others (BB)
* Non-submission workshop (BB)

as well as many bugfixes and compatibility with Moodle 3.1.5. (And probably some other stuff I've forgotten.)

Please check the cldev-teameval branch on this repository for individual commits.